chore: opitimize directory
