+++ packages/rax-miniapp-runtime-webpack-plugin/src/utils/addFileToCompilation.js
@@ -1,0 +1,7 @@
+// Add file to compilation
+module.exports = function(compilation, filename, content, target) {
+  compilation.assets[`${target}/${filename}`] = {
+    source: () => content,
+    size: () => Buffer.from(content).length
+  };
+};
